♻️ refactor: refactor main and support text input
